chore: rename tool.ts to util.ts
